SOALR SYSTEM BABY

+++ physics.h
@@ -7,6 +7,6 @@
 
 inline constexpr double G = 6.67430e-11; // m^3 kg^-1 s^-2
 
-void nextPosition(float dt, PlanetGroup& group);
+void nextPosition(double long dt, PlanetGroup& group);
 
 #endif
