+++ planet.cpp
@@ -4,7 +4,7 @@
 #include "planet.h"
 #include <SFML/Graphics.hpp>
 
-Planet::Planet(const float mass, const float radius, const sf::Vector2f velocity, const sf::Vector2f position, const sf::Color color, const bool isStatic)
+Planet::Planet(const double long mass, const double long radius, const sf::Vector2f velocity, const sf::Vector2f position, const sf::Color color, const bool isStatic)
     : mass(mass), radius(radius), velocity(velocity), position(position), color(color), isStatic(isStatic){
         shape.setRadius(radius * RADIUS_SCALE);
         shape.setFillColor(color);
